Added filter prior to encoding giving BDR gains on certain sequences like KristenAndSara.
It tries to eliminate noisy static areas of the
video so bits aren't spent in encoding this noise.  Ideally the encoder
should be to handle this and this prefilter should be removed when it
no longer gives an improvement.  Must be enabled by a #define near the
top of encoder_example.c.